fix: remove serious bug from rating
After a deep dive into the rating function, we realized it was not working correctly. This commit has made sure the localStorage and database both update when giving a rating. There are still some issues with displaying it, but all logic is sound. You currently need to refresh the site to see the changes.

Co-authored-by: Otto Fearnley ottohf@stud.ntnu.no

Issue: #44