feat: host the final version of the site on the vm
Do not merge this branch.

This commit contains changes to build the project and out both backend and frontend on the VM. It serves no purpose when running locally, and should be disregarded. The deletions of tests was done because we were struggling to build, and wanted to avoid issues with testing.

Co-authored-by: Hanna Lunne hannalun@stud.ntnu.no

Co-authored-by: Ola Haldor Hådi olahh@stud.ntnu.no

Issue: #58